Implement cdevcontainer CLI tool and improve project structure (#3)
* Added notes to README about support for AMAZONQ even if the project does not depend AWS.
Added an ENV VAR to the devcontainer.json so that other development project build task that may automate the install of tools via ASDF can be aware it is running in a devcontainer and already had this completed so the task can skip asdf automation.

* Wrapped all env config logic into a CLI tool for better usability and maintainability.
Setup support for concurrent devcontainers to support 1 to many clients and 1 to mant project.
Setup a means to reuse project templates for devcontainers.
Added an environment variable to the devcontainer to enable other utilities that may be deployed on the container to determine its running in a devcontainer or not.

* Moved supporting docs out of repo root

* fixed amazonq

* Added info to README about when and how to rebuild the devcontainer

* fixed linting and some nit test

* fix unit test and linting

* got unit test coverage up to 91%

* fix install of asdf tools and remove launch-cli

* fix linting

* removed 502 ignore in flake

* updated isort dep